further clean up and fix bugs
