+++ model/common/src/icon4py/model/common/metrics/metric_fields.py
@@ -6,6 +6,7 @@
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
 from types import ModuleType
+from typing import Final
 
 import gt4py.next as gtx
 import numpy as np
@@ -29,7 +30,7 @@ from gt4py.next import (
 )
 from gt4py.next.ffront.experimental import concat_where
 
-from icon4py.model.common import dimension as dims, field_type_aliases as fa, option_groups
+from icon4py.model.common import dimension as dims, field_type_aliases as fa, model_options
 from icon4py.model.common.dimension import (
     C2E,
     C2E2C,
@@ -56,6 +57,9 @@ from icon4py.model.common.utils import data_allocation as data_alloc
 """
 Contains metric fields calculations for the vertical grid, ported from mo_vertical_grid.f90.
 """
+
+
+rayleigh_damping_options: Final = model_options.RayleighType()
 
 
 # TODO(@nfarabullini): ddqz_z_half vertical dimension is khalf, use K2KHalf once merged for z_ifc and z_mc
@@ -225,13 +229,13 @@ def _compute_rayleigh_w(
     rayleigh_w = broadcast(0.0, (dims.KDim,))
     z_sin_diff = maximum(0.0, vct_a - damping_height)
     z_tanh_diff = vct_a_1 - vct_a  # vct_a(1) - vct_a
-    if rayleigh_type == option_groups.RayleighType.CLASSIC:
+    if rayleigh_type == rayleigh_damping_options.CLASSIC:
         rayleigh_w = (
             rayleigh_coeff
             * (sin(pi_const / 2.0 * z_sin_diff / maximum(0.001, vct_a_1 - damping_height))) ** 2
         )
 
-    elif rayleigh_type == option_groups.RayleighType.KLEMP:
+    elif rayleigh_type == rayleigh_damping_options.KLEMP:
         rayleigh_w = rayleigh_coeff * (
             1.0 - tanh(3.8 * z_tanh_diff / maximum(0.000001, vct_a_1 - damping_height))
         )
