+++ model/common/src/icon4py/model/common/model_options.py
@@ -6,7 +6,6 @@
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
 
-
 from gt4py.eve.utils import FrozenNamespace
